Add the interpreters for HELICS3 binary format to HELICS 2 (#2088)
* Add the interpreters for HELICS3 binary format to HELICS 2 and some tests to check some aspects of it.

* fix some compiler warnings

* fix a few more warnings

* Automated formatting of repo files (#2094)


* fix cpplint issue

+++ tests/helics/core/ActionMessage-tests.cpp
@@ -435,7 +435,7 @@ TEST(ActionMessage, check_json_packetization)
     auto cmdString = cmd.packetize_json();
     EXPECT_GE(cmdStringNormal.size() + 6, cmdString.size());
     helics::ActionMessage cmd2;
-    auto res = cmd2.depacketize(cmdString.data(), cmdString.size());
+    auto res = cmd2.depacketize(cmdString.data(), static_cast<int>(cmdString.size()));
     EXPECT_EQ(res, static_cast<int>(cmdString.size()));
     EXPECT_TRUE(cmd.action() == cmd2.action());
     EXPECT_EQ(cmd.actionTime, cmd2.actionTime);
